kpatch: make it easier to tune the number of attempts to load the patch
... and the interval between the retries.

If activeness safety check fails and the patch fails to load, kpatch waits
for 2 seconds and then retries loading of that patch.

It may be needed to change the number of retries and the interval
between them in some cases, e.g. during stress testing of patches, etc.

Make it a bit easier by keeping these values in the variables close to
the beginning of the script.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>