kpatch-build: fix two parent matches error
when I make a patch for a module, cause the error as follow:
ERROR: two parent matches for hello.o. Check /tmp/kpatch-build.log for more details.

The Makefile is as follow:
	obj-m += m_hello.o
	m_hello-y = hello.o
	default:
        	$(MAKE) -C /lib/modules/4.4.21-69-default/build M=$(shell pwd) modules
	clean:
        	$(MAKE) -C /lib/modules/4.4.21-69-default/build M=$(shell pwd) clean

and than cmd "grep -l hello.o ./.*.cmd | grep -Fv hello.o" would get two parant
	.m_hello.ko.cmd
	.m_hello.o.cmd
but the ".m_hello.ko.cmd" is the parant for "m_hello.o", and ".m_hello.o.cmd" is the
parant for "hello.o".

because the "hello.o" is a substring of "m_hello.o", which will cause the "m_hello.o"
matched for the "hello.o".

Signed-off-by: chenzefeng <chenzefeng2@huawei.com>