kpatch-build: use whole word filename matching in find_parent_obj()
Building a kpatch for a module with this Makefile:

The Makefile is as follow:
	obj-m += m_hello.o
	m_hello-y = hello.o
	default:
        	$(MAKE) -C /lib/modules/4.4.21-69-default/build M=$(shell pwd) modules
	clean:
        	$(MAKE) -C /lib/modules/4.4.21-69-default/build M=$(shell pwd) clean

results in kpatch-build "ERROR: two parent matches for hello.o".

The problem is that find_parent_obj() looks for filenames like so:

	% grep -l hello.o ./.*.cmd | grep -Fv hello.o
	.m_hello.ko.cmd
	.m_hello.o.cmd

where .m_hello.ko.cmd is the parant for m_hello.o, and .m_hello.o.cmd is the
parant for hello.o, but because the "hello.o" is a substring of "m_hello.o",
it will cause "m_hello.o" to be matched for the "hello.o" as well.

Fix this by using grep's -w|--word-regexp option to force it to match
whole words instead of substrings.

Signed-off-by: chenzefeng <chenzefeng2@huawei.com>