kpatch: Fix all module unloading
Some versions of Linux livepatching require patches to be unloaded
in the opposite order than the order they were loaded. Currently
kpatch simply goes through the list of sysfs livepatch directories
and tries to disable them one by one. This makes the following fail:

	$ kpatch load a.ko
	$ kpatch load b.ko
	$ kpatch unload --all
	kpatch: failed to unload module b

Since kpatch is not bound to a version of Linux, cater for those version
that do not support random unloading order, even for patches that do not
depend on one another.

Signed-off-by: Julien Thierry <jthierry@redhat.com>