kpatch-build: Create module symbol version file once
When patching an OOT module, the symbol version file is obtained by
combining the file from the module build and the Module.symvers file
provided with kernel headers. This is done for each modified .o in the
OOT build.

Create the final Module.symvers file once for the whole OOT module.

Signed-off-by: Julien Thierry <jthierry@redhat.com>